fix(web): repair iframe bridge factory wiring
