+++ mobile/lib/services/web_iframe_nostr_bridge_binding.dart
@@ -1,0 +1,16 @@
+typedef WebIframeBridgeMessageHandler =
+    Future<void> Function({
+      required String origin,
+      required dynamic data,
+      required void Function(dynamic message, String targetOrigin) postReply,
+    });
+
+abstract class WebIframeNostrBridgeBinding {
+  void start({
+    required WebIframeBridgeMessageHandler onMessage,
+    required void Function(dynamic message, String targetOrigin)? Function()
+    getPostReplyOverride,
+  });
+
+  void stop();
+}
